+++ vite.config.js
@@ -1,0 +1,39 @@
+import { defineConfig } from 'vite'
+import tailwindcss from '@tailwindcss/vite'
+import laravel from 'laravel-vite-plugin'
+import { wordpressPlugin, wordpressThemeJson } from '@roots/vite-plugin'
+
+if (!process.env.APP_URL) {
+  process.env.APP_URL = 'http://localhost:8080'
+}
+
+export default defineConfig({
+  base: '/wp-content/themes/walkridge/public/build/',
+  plugins: [
+    tailwindcss(),
+    laravel({
+      input: [
+        'resources/css/app.css',
+        'resources/js/app.js',
+        'resources/css/editor.css',
+        'resources/js/editor.js',
+      ],
+      refresh: true,
+      assets: ['resources/images/**', 'resources/fonts/**'],
+    }),
+    wordpressPlugin(),
+    wordpressThemeJson({
+      disableTailwindColors: true,
+      disableTailwindFonts: true,
+      disableTailwindFontSizes: true,
+    }),
+  ],
+  resolve: {
+    alias: {
+      '@scripts': '/resources/js',
+      '@styles': '/resources/css',
+      '@fonts': '/resources/fonts',
+      '@images': '/resources/images',
+    },
+  },
+})